third_party/blink/web_tests/transforms/3d/hit-testing/negative-zoffset-hit-test-expected.txt

Checks that hit testing is correct when a transformed element has negative z-offset

Element at 50, 50 has id "container": PASS
Element at 60, 60 has id "container": PASS
Element at 70, 70 has id "target": PASS
Element at 80, 80 has id "target": PASS
Element at 90, 90 has id "target": PASS
Element at 100, 100 has id "target": PASS
Element at 150, 150 has id "target": PASS
Element at 180, 180 has id "container": PASS
